% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAScoreMatrix.R
\name{as.data.frame.PCAScoreMatrix}
\alias{as.data.frame.PCAScoreMatrix}
\title{Coerece a PCAScoreMatrix into data.frame}
\usage{
\method{as.data.frame}{PCAScoreMatrix}(x, row.names = NULL,
  optional = FALSE, ...)
}
\arguments{
\item{x}{A \code{PCAScoreMatrix} S3 object.}

\item{row.names}{See \code{as.data.frame}.}

\item{optional}{See \code{as.data.frame}.}

\item{...}{See \code{as.data.frame}}
}
\value{
A data.frame consisting of the score matrix
}
\description{
Coerece a PCAScoreMatrix into data.frame
}
\examples{
myPCmat <- PCAScoreMatrix(matrix(rnorm(15),ncol=3), c(0.25, 0.15, 0.1))
as.matrix(myPCmat)
}
\seealso{
\code{\link{as.data.frame}}
}
