% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAScoreMatrix.R
\name{as.matrix.PCAScoreMatrix}
\alias{as.matrix.PCAScoreMatrix}
\title{Coerece a PCAScoreMatrix into score matrix}
\usage{
\method{as.matrix}{PCAScoreMatrix}(x, ...)
}
\arguments{
\item{x}{A \code{PCAScoreMatrix} S3 object}

\item{...}{Currently ignored}
}
\value{
A numeric matrix, the score matrix
}
\description{
Coerece a PCAScoreMatrix into score matrix
}
\examples{
myPCmat <- PCAScoreMatrix(matrix(rnorm(15),ncol=3), c(0.25, 0.15, 0.1))
as.matrix(myPCmat)
}
