% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{blackyellow}
\alias{blackyellow}
\alias{yellowblack}
\alias{whiteblue}
\alias{whitered}
\alias{blackred}
\alias{blackgreen}
\alias{whiteblack}
\alias{blackwhite}
\alias{twocolor.panels}
\alias{display.twocolor.panels}
\title{Two-color panels}
\usage{
blackyellow(n)
}
\arguments{
\item{n}{Number of colors needed}

\item{nc}{Number of color columns}
}
\value{
Character vector of length \code{n} coding colors
}
\description{
Precompiled two-color panels for visualization purposes
}
\details{
\code{twocolor.panels} returns all available two-color panels.
\code{display.twocolor.panels} display available two-color panels.

For the rest see function definitions.
}
\examples{

display.twocolor.panels()

}
\seealso{
\code{\link{royalbluered}} for three-color systems
}
\author{
Jitao David Zhang
}
