% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compactPar}
\alias{compactPar}
\title{Compact par setting}
\usage{
compactPar(mar = c(3, 3, 1.5, 1.5), mgp = c(2, 1, 0), ...)
}
\arguments{
\item{mar}{marginal option passed to \code{par}}

\item{mgp}{margin line option passed to \code{par}}

\item{...}{other parameters passed to \code{par}}
}
\value{
A named list of the previous \code{par} settings (invisibly),
  as returned by \code{\link{par}}.
}
\description{
For compact figures
}
\examples{

  compactPar()
  plot(1:4)

}
\seealso{
\code{\link{par}}
}
\author{
Jitao David Zhang
}
