% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse.R
\name{degree2radian}
\alias{degree2radian}
\title{Convert degree to radian values}
\usage{
degree2radian(x)
}
\arguments{
\item{x}{Degree value}
}
\value{
Radian value
}
\description{
Convert degree to radian values
}
\examples{

degree2radian(90)
degree2radian(-225)
}
