% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{display.colorpanels}
\alias{display.colorpanels}
\title{Display color panels}
\usage{
display.colorpanels(panel.names, nc)
}
\arguments{
\item{panel.names}{A vector of character strings, panels to be visualized}

\item{nc}{Number of color columns}
}
\value{
Side effect (visuzalization is used)
}
\description{
Display color panels
}
\examples{
display.colorpanels(threecolor.panels(), 6)
}
\author{
Jitao David Zhang
}
