% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{display.threecolor.panels}
\alias{display.threecolor.panels}
\title{Display three-color panels}
\usage{
display.threecolor.panels(nc = 20)
}
\arguments{
\item{nc}{Number of columns}
}
\value{
Side effect is used
}
\description{
Display three-color panels
}
\examples{

display.threecolor.panels()
}
