% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{display.twocolor.panels}
\alias{display.twocolor.panels}
\title{Display two-color panels}
\usage{
display.twocolor.panels(nc = 20)
}
\arguments{
\item{nc}{Number of columns}
}
\value{
Side effect is used
}
\description{
Display two-color panels
}
\examples{

display.twocolor.panels()
}
