% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R, R/pcaPlots.R
\name{expVarLabel}
\alias{expVarLabel}
\title{Generic function expVarLabel to generate a label of explained variance from prcomp and PCAScoreMatrix objects}
\usage{
expVarLabel(prcomp, choices = 1, compact = FALSE)

expVarLabel(prcomp, choices = 1, compact = FALSE)
}
\arguments{
\item{prcomp}{An object of prcomp}

\item{choices}{The choice(s) of principal components}

\item{compact}{Logical, whether the label should be compact. See examples.}
}
\value{
Character string vector of the same length as \code{choices}
}
\description{
Generic function expVarLabel to generate a label of explained variance from prcomp and PCAScoreMatrix objects

Label of explained variance of the chosen principal component(s)
}
\examples{
testData <- matrix(rnorm(100), 10, 10)
testPrcomp <- prcomp(testData)
expVarLabel(testPrcomp, 1:2)
expVarLabel(testPrcomp, 1:2, compact=TRUE)
}
