% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcol.R
\name{fcbase<-}
\alias{fcbase<-}
\title{Replace base colors of a fcol object with a different value}
\usage{
fcbase(fcol) <- value
}
\arguments{
\item{fcol}{A fcol object, likely constructed by \code{\link{fcol}}}

\item{value}{A character vector, indicating the new values}
}
\value{
A new \code{fcol} object
}
\description{
Replace base colors of a fcol object with a different value
}
\examples{
fc <- fcol(c("lightblue", "orange", "lightblue"), base=c("orange", "lightblue"))
fcbase(fc)
fcbase(fc)[1] <- "red"
print(fc)
fcbase(fc) <- c("gray", "darkblue")
fc
}
