% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcol.R
\name{fcbase}
\alias{fcbase}
\title{Return base colors of a fcol object}
\usage{
fcbase(fcol)
}
\arguments{
\item{fcol}{A fcol object, likely constructed by \code{\link{fcol}}}
}
\value{
A character vector, representing base colors
}
\description{
Return base colors of a fcol object
}
\examples{
fc <- fcol(c("lightblue", "orange", "lightblue"), base=c("orange", "lightblue"))
fcbase(fc)
}
