% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribiosPlot-package.R
\docType{methods}
\name{fcol-methods}
\alias{fcol-methods}
\alias{fcol,character,character-method}
\title{Build fcol objects}
\description{
Build fcol objects
}
\section{Methods}{
 \describe{

\item{list("signature(object = \"character\", base = \"character\")")}{
Object should be a vector of colors, and base should be the unique elements.
} }
}

