% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcol.R
\name{fcol}
\alias{fcol}
\title{Construct a fcol object}
\usage{
fcol(colors, base)
}
\arguments{
\item{colors}{A character vector, containing colors}

\item{base}{A character vector, containing base colors}
}
\value{
A S3 class known as \code{fcol}, containing colors as vector and base colors in the attribute \code{fcbase}
}
\description{
Construct a fcol object
}
\examples{
fcol(c("lightblue", "orange", "lightblue"), base=c("orange", "lightblue"))
}
