% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expVar.R
\name{getExpVarLabel}
\alias{getExpVarLabel}
\title{Helper function to print PC and explained variances}
\usage{
getExpVarLabel(ev, choices, compact = FALSE)
}
\arguments{
\item{ev}{A numeric vector of explained variances}

\item{choices}{An integer vector to indicate which PCs to be returned. If \code{NULL} or \code{NA} or missing, all elements are returned.}

\item{compact}{Logical, either a \code{compact} label is returned, see examples.}
}
\description{
Helper function to print PC and explained variances
}
