% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{intRange}
\alias{intRange}
\title{Return a range defined by integers}
\usage{
intRange(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector}

\item{na.rm}{Logical, whether NA should be removed}
}
\value{
A vector of integers of length 2.
}
\description{
The function is similar to \code{\link{range}} but returns integer ranges
that are just outside the real range: i.e. the floor of the left range and
the ceiling of the right range.
}
\examples{

intRange(rnorm(100))

}
\author{
Jitao David Zhang <jitao_david.zhang@roche.com>
}
