% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isOdd}
\alias{isOdd}
\alias{isEven}
\title{Whether an integer is odd (or even)}
\usage{
isOdd(x)

isEven(x)
}
\arguments{
\item{x}{An integer.}
}
\value{
Logical, whether the input number is odd or even.

\code{isOdd} and \code{isEven} returns whether an integer is odd or even,
respectively.
}
\description{
Whether an integer is odd (or even)
}
\examples{
isOdd(3)
isEven(4)
}
