% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nonNull}
\alias{nonNull}
\title{Make sure that x is assigned a reasonable value}
\usage{
nonNull(x, default, length = NULL, defaultNULL.ok = FALSE)
}
\arguments{
\item{x}{Any vector}

\item{default}{A default value}

\item{length}{Desired length}

\item{defaultNULL.ok}{Logical, whether the default can be \code{NULL} or not}
}
\value{
non-null values
}
\description{
Make sure that x is assigned a reasonable value
}
