% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs.R
\name{panel.cor}
\alias{panel.cor}
\title{Correlation panel for pairs}
\usage{
panel.cor(x, y, digits = 2, prefix = "", cex.cor, ...)
}
\arguments{
\item{x}{A numeric vector}

\item{y}{A numeric vector, must be of the same length as \code{x}.}

\item{digits}{Integer, number of digits to show}

\item{prefix}{Prefix of the label}

\item{cex.cor}{Numeric, if missing, automatically guessed}

\item{...}{Passed to \code{cor}.

This function can be used with \code{pairs} to display correlations.}
}
\description{
Correlation panel for pairs
}
\seealso{
\code{\link[graphics]{pairs}}.
}
