% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs.R
\name{panel.lmSmooth}
\alias{panel.lmSmooth}
\title{Correlation panel}
\usage{
panel.lmSmooth(
  x,
  y,
  col = par("col"),
  bg = NA,
  pch = par("pch"),
  cex = 0.8,
  method = "spearman",
  use = "complete",
  ...
)
}
\arguments{
\item{x}{A numeric vector}

\item{y}{A numeric vector, must be of the same length as \code{x}.}

\item{col}{Color}

\item{bg}{Background}

\item{pch}{Ponit symbol}

\item{cex}{Font size}

\item{method}{Correlation method, passed to \code{\link[stats]{cor}}}

\item{use}{passed to \code{cor}}

\item{...}{Passed to \code{\link[graphics]{panel.smooth}}

This function can be used with \code{pairs} to display correlations.}
}
\description{
Correlation panel
}
\seealso{
\code{\link[graphics]{pairs}}.
}
