% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{PCAScoreMatrix-class}
\alias{PCAScoreMatrix-class}
\title{A simple data structure to hold PCA scores in numeric matrix and explained variance in dedicated slots}
\description{
A simple data structure to hold PCA scores in numeric matrix and explained variance in dedicated slots
}
