% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{PCAScoreMatrix}
\alias{PCAScoreMatrix}
\title{Construct a PCAScoreMatrix object}
\usage{
PCAScoreMatrix(scoreMatrix, expvar)
}
\arguments{
\item{scoreMatrix}{Numeric matrix, objects in rows and PCs in columns}

\item{expvar}{Numeric vector, length must equal the number of columns of \code{scoreMatrix}, explained variance by respective PCs

This function is usually not called by the end user; instead, it is used by the function \code{\link{pcaScores}}}
}
\description{
Construct a PCAScoreMatrix object
}
\examples{
PCAScoreMatrix(matrix(rnorm(15),ncol=3), c(0.25, 0.15, 0.1))
}
