% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaPlots.R
\name{pcaScores}
\alias{pcaScores}
\title{Retrieve PCA scores from prcomp objects}
\usage{
pcaScores(x, choices = c(1, 2), offset)
}
\arguments{
\item{x}{An object of prcomp}

\item{choices}{indices of principal components}

\item{offset}{either one or more rows's names in the loading matrix, or indices, or a logical vector. The average loading of the rows specified by offset is set to zero.}
}
\description{
Retrieve PCA scores from prcomp objects
}
\examples{
testMatrix <- matrix(rnorm(1000), nrow=10)
testPCA <- prcomp(testMatrix)
testPCAscores <- pcaScores(testPCA)

testPCAscores.withOffset <- pcaScores(testPCA, offset=c(1,3,5))
## notice the average of offset-rows are near zero
colMeans(testPCAscores.withOffset[c(1,3,5),])
}

