% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{pcLabels,PCAScoreMatrix-method}
\alias{pcLabels,PCAScoreMatrix-method}
\title{Labels of principal components}
\usage{
\S4method{pcLabels}{PCAScoreMatrix}(object, variant = c("compact", "full"))
}
\description{
Labels of principal components
}
