% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaPlots.R
\name{plotPCA}
\alias{plotPCA}
\title{S3 method plotPCA}
\usage{
plotPCA(x, choices, ...)
}
\arguments{
\item{x}{A prcomp object}

\item{choices}{Integer index, choices to plot}

\item{...}{Other parameters}
}
\description{
S3 method plotPCA
}
