% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaPlots.R
\name{plotPCAloading}
\alias{plotPCAloading}
\title{Plot PCA loading}
\usage{
plotPCAloading(
  loadings,
  x = 1L,
  y = 2L,
  circle = FALSE,
  title = "",
  subtitle = "",
  ...
)
}
\arguments{
\item{loadings}{PCA loadings}

\item{x}{Integer, which loading is to be visualized on the X axis?}

\item{y}{Integer, the loading to be visualized on the X axis.}

\item{circle}{Logical, whether draw circle or not}

\item{title}{Character string}

\item{subtitle}{Character string}

\item{...}{Passed to \code{\link{plot}}}
}
\value{
No return value, called for side effects (plotting).
}
\description{
Plot PCA loading
}
