% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcol.R
\name{print.fcol}
\alias{print.fcol}
\title{Print a fcol object}
\usage{
\method{print}{fcol}(x, ...)
}
\arguments{
\item{x}{A fcol object, likely constructed by \code{\link{fcol}}}

\item{...}{Not used now}
}
\value{
The input \code{x}, invisibly.
}
\description{
Print a fcol object
}
\examples{
fc <- fcol(c("lightblue", "orange", "lightblue"), base=c("orange", "lightblue"))
fc
}
