% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaPlots.R
\name{printExpVar}
\alias{printExpVar}
\title{Print explained variance of the chosen principal component(s)}
\usage{
printExpVar(prcomp, choices = 1)
}
\arguments{
\item{prcomp}{An object of prcomp}

\item{choices}{The choice(s) of principal components}
}
\description{
Print explained variance of the chosen principal component(s)
}
\examples{
testData <- matrix(rnorm(100), 10, 10)
testPrcomp <- prcomp(testData)
printExpVar(testPrcomp, 1:2)
}
