% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse.R
\name{radian2degree}
\alias{radian2degree}
\title{Convert radian to degree values}
\usage{
radian2degree(x)
}
\arguments{
\item{x}{Radian value}
}
\value{
Degree value
}
\description{
Convert radian to degree values
}
\examples{
radian2degree(2*pi)
radian2degree(-.5*pi)
}

