% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{royalbluered}
\alias{royalbluered}
\alias{royalredblue}
\alias{royalbluegrayred}
\alias{royalredgrayblue}
\alias{turyeb}
\alias{redgreen}
\alias{greenred}
\alias{bluered}
\alias{redblue}
\alias{redblackblue}
\alias{cyanblackyellow}
\alias{yellowblackcyan}
\alias{blueblackred}
\alias{blackredyellow}
\alias{blackgoldred}
\alias{whiteblueblackheat}
\alias{heat}
\alias{magentayellow}
\alias{yellowmagenta}
\alias{threecolor.panels}
\alias{display.threecolor.panels}
\title{Three-color panels}
\usage{
royalbluered(n)
}
\arguments{
\item{n}{Number of colors needed}

\item{nc}{Number of color columns}
}
\value{
Character vector of length \code{n} coding colors
}
\description{
Precompiled three-color panels for visualization purposes
}
\details{
\code{threecolor.panels} returns all available three-color panels.
\code{display.threecolor.panels} display available three-color panels.

For the rest see function definitions
}
\examples{

display.threecolor.panels()

}
\seealso{
\code{\link{blackyellow}} for two-color systems
}
\author{
Jitao David Zhang
}
