% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{show,PCAScoreMatrix-method}
\alias{show,PCAScoreMatrix-method}
\title{Show PCAScoreMatrix object}
\usage{
\S4method{show}{PCAScoreMatrix}(object)
}
\arguments{
\item{object}{A PCAScoreMatrix object}
}
\description{
Show PCAScoreMatrix object
}
