% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAScoreMatrix.R
\name{[.PCAScoreMatrix}
\alias{[.PCAScoreMatrix}
\title{Subsetting PCAScoreMatrix while keeping the expVar attribute}
\usage{
\method{[}{PCAScoreMatrix}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{A \code{PCAScoreMatrix} object}

\item{i}{Integer or logical vector, subsetting rows}

\item{j}{Integer or logical vector, subsetting columns}

\item{...}{Not used}

\item{drop}{Logical, whether to drop dimensions if only one column is left}
}
\value{
A \code{PCAScoreMatrix} object
}
\description{
Subsetting PCAScoreMatrix while keeping the expVar attribute
}
