% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{symrange}
\alias{symrange}
\title{Return a symmetric range}
\usage{
symrange(x, mid = 0)
}
\arguments{
\item{x}{A numeric vector}

\item{mid}{Number, the mid point}
}
\value{
A vector of two numbers, a symmetric range with \code{mid} in the middle
}
\description{
Return a symmetric range
}
