% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{threecolor.panels}
\alias{threecolor.panels}
\title{Return available three-color panels}
\usage{
threecolor.panels()
}
\value{
A vector of character strings
}
\description{
Return available three-color panels
}
