% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{brewer.pal.factor}
\alias{brewer.pal.factor}
\alias{brewer.pal.factorLevels}
\title{Build brewer.pal colors from factor (Deprecated)}
\usage{
brewer.pal.factor(factor, name = "Greys")
}
\arguments{
\item{factor}{A factor vector}

\item{name}{Color panel name to be passed to \code{brewer.pal}}
}
\value{
Named HTML RGB colors.
}
\description{
The functionality has been replaced by \code{fcbrewer}. The functions will
be removed in the future release.
}
\details{
The function is useful to build named RGB color values from factors.
\code{brewer.pal.factor} return a color-HTML-string vector as the same
length of the input factor vector, which is named by the input factor as
well. \code{brewer.pal.factorLevels} returns a color vector of the length of
the factor level, and the colors are named by the levels. See examples
below.

From version 1.1-16, the color palette is automatically reduced/expanded
when the levels of input factors underlies or exceeds the minimum and
maximum colors. See example below.
}
\examples{

\donttest{
myFac <- factor(c("HSV", "BVB", "FCB", "HSV", "BVB", "HSV"))
brewer.pal.factor(myFac, name="Set1")

myLongFac <- factor(paste("Sample", 1:20))
brewer.pal.factor(myLongFac, name="Set1")

myShortFac <- factor(paste("Sample", 1:2))
brewer.pal.factor(myShortFac, name="Set1")
}

}
\author{
Jitao David Zhang <jitao_david.zhang@roche.com>
}
