% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cascadePlot.R
\name{cascadeOrder}
\alias{cascadeOrder}
\title{Order rows of a matrix in the cascade order}
\usage{
cascadeOrder(matrix, dichotomy = c("mean", "median", "maxabs"))
}
\arguments{
\item{matrix}{A numeric matrix}

\item{dichotomy}{How are the rows divided into two? By mean value, median value, or by maximal abs value}
}
\description{
The 'cascade order' is defined by (1) rows are divided into two groups by the condition given by 'dichotomy'
(2) the positive and negative rows are ordered respectively so that rows reaching its absolute maximal values
in column n are ordered prior to rows reaching reaching its absolute maximal values in columns n+1, where n can
be from 1 to column number minus one.
}
\examples{
myMatrix <- matrix(c(1,2,3, 4,6,5, 8,7,9,3,4,8,9,1,2,-1,-2,-3, -4, -2, -1, -5, -7, -4), byrow=TRUE, ncol=3)
rownames(myMatrix) <- sprintf("Gene\%d", 1:nrow(myMatrix))
biosHeatmap(myMatrix, Colv=FALSE, Rowv=FALSE, dendrogram="none")
myOrder <- cascadeOrder(myMatrix)
myOrderedMatrix <- myMatrix[myOrder,]
biosHeatmap(myOrderedMatrix, Colv=FALSE, Rowv=FALSE, dendrogram="none")
}
