% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R, R/pcaPlots.R
\name{expVar}
\alias{expVar}
\title{S3 function expVar to extract explained variance from prcomp and PCAScoreMatrix objects}
\usage{
expVar(prcomp, choices = seq(along = prcomp$sdev))

expVar(prcomp, choices = seq(along = prcomp$sdev))
}
\arguments{
\item{prcomp}{An object of prcomp}

\item{choices}{Either missing, or an integer vector of indices, indicating which PCs should be returned}

\item{x}{A \code{prcomp} or \code{PCAScoreMatrix} object}

\item{choices}{Indices of principal components. By default explained variance of all components are returned}
}
\value{
A numeric vector of variance explained
}
\description{
S3 function expVar to extract explained variance from prcomp and PCAScoreMatrix objects

Retrieve percentage of variance explained by principal components
}
\examples{
myMat <- matrix(rnorm(100), ncol=10)
myPrcomp <- prcomp(myMat)
myPcaScoreMatrix <- pcaScores(myPrcomp, choices=NULL) 
expVar(myPrcomp)
expVar(myPcaScoreMatrix)

expVar(myPrcomp, 1:5)
expVar(myPcaScoreMatrix, 1:5)
testData <- matrix(rnorm(100), 10, 10)
testPrcomp <- prcomp(testData)
expVar(testPrcomp)
expVar(testPrcomp, 1:3)
}
