% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaPlots.R
\name{expVar}
\alias{expVar}
\title{Retrieve percentage of variance explained by principal components}
\usage{
expVar(prcomp, choices = seq(along = prcomp$sdev))
}
\arguments{
\item{prcomp}{An object of prcomp}

\item{choices}{Indices of principal components. By default explained variance of all components are returned}
}
\description{
Retrieve percentage of variance explained by principal components
}
\examples{
testData <- matrix(rnorm(100), 10, 10)
testPrcomp <- prcomp(testData)
expVar(testPrcomp)
expVar(testPrcomp, 1:3)
}
