% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{expVarLabel}
\alias{expVarLabel}
\title{Generic function expVarLabel to generate a label of explained variance from prcomp and PCAScoreMatrix objects}
\usage{
expVarLabel(x, choices, compact)
}
\arguments{
\item{x}{\code{prcomp} or \code{PCAScoreMatrix} Object}

\item{choices}{Integer indices of which PCs to be returned}

\item{compact}{Logical, whether a compact format is returned, see example}
}
\description{
Generic function expVarLabel to generate a label of explained variance from prcomp and PCAScoreMatrix objects
}
