% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribiosPlot-package.R
\docType{class}
\name{fcol-class}
\alias{fcol-class}
\alias{fcol}
\alias{fcbase}
\alias{fcbase-methods}
\alias{fcbase,fcol-method}
\alias{show,fcol-method}
\title{Factor-matching colors}
\description{
This object represents a vector of colors (in characters), together with a
vector of 'base colors', i.e. the elementary colors used in the vector. This
data structure is useful to represent factor using colors since the base
colors are mapped to the levels of the factor.
}
\section{Objects from the Class}{
 Objects can be created by calls of the
function \code{fcol}. However, the users should not directly call this
function. Instead, \code{\link{fcbrewer}} should be called to generate a
factor-matching color vector. \code{fcbrewer} calls \code{fcol} internally.
}

\examples{

  testFac <- gl(4,2)
  testCol <- fcbrewer(testFac, panel="Set1")

  print(testCol)
  fcbase(testCol)
  as.character(testCol)

}
\author{
Jitao David Zhang <jitao_david.zhang@roche.com>
}
\keyword{classes}
