\name{fcol-methods}
\docType{methods}
\alias{fcol-methods}
\alias{fcol,character,character-method}
\title{ Build fcol objects}
\description{
  Build fcol objects
}
\section{Methods}{
\describe{

\item{\code{signature(object = "character", base = "character")}}{
  Object should be a vector of colors, and base should be the unique
  elements.
}
}}
