% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggSmoothScatter.R
\name{ggSmoothScatter}
\alias{ggSmoothScatter}
\title{Mimicking the graphics::smoothScatter behaviour for GGally::ggpairs}
\usage{
ggSmoothScatter(
  data,
  mapping,
  colours = colorRampPalette(c("white", blues9[5:9], "black"))(256),
  xlim = NULL,
  ylim = NULL,
  ...
)
}
\arguments{
\item{data}{Data to be visualized, normally not directly set by the user}

\item{mapping}{Data mapping, normally not directly set by the user}

\item{colours}{Colours to be used}

\item{xlim}{NULL or a vector of two numbers}

\item{ylim}{NULL or a vector of two numbers}

\item{...}{Other parameters passed to stat_density_2d

#' @note 
So far the outliers are not plotted, to be done later}
}
\description{
Mimicking the graphics::smoothScatter behaviour for GGally::ggpairs
}
