% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist.R
\name{histMat}
\alias{histMat}
\title{Make histograms for matrix}
\usage{
histMat(
  mat,
  linesOpt = list(lwd = NULL, col = NULL, lty = NULL, type = NULL, pch = NULL),
  main = NULL,
  xlab = NULL,
  xlim = NULL,
  ...
)
}
\arguments{
\item{mat}{A numerical matrix}

\item{linesOpt}{Line options}

\item{main}{Title text}

\item{xlab}{Xlab}

\item{xlim}{Xlim}

\item{\dots}{Other parameters passed to \code{hist}}
}
\description{
Make histograms for matrix
}
\examples{

testMat <- matrix(rnorm(1000), nrow=100)
histMat(testMat)

}
\seealso{
\code{\link{hist}}
}
\author{
Jitao David Zhang <jitao_david.zhang@roche.com>
}
