% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribiosPlot-package.R
\docType{package}
\name{ribiosPlot-package}
\alias{ribiosPlot}
\alias{ribiosPlot-package}
\title{ribiosPlot: Plotting Module of Ribios for Visualization}
\description{
Provides data structures and functions for data transformation and visualization in computational biology. Includes heatmaps, color schemes, PCA plots, and various plotting utilities for drug discovery.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bedapub/ribiosPlot}
  \item Report bugs at \url{https://github.com/bedapub/ribiosPlot/issues}
}

}
\author{
\strong{Maintainer}: Jitao David Zhang \email{jitao_david.zhang@roche.com} (\href{https://orcid.org/0000-0002-3085-0909}{ORCID}) [contributor]

Other contributors:
\itemize{
  \item F.Hoffmann-La Roche AG [copyright holder]
}

}
\keyword{internal}
