% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribiosPlot-package.R
\docType{package}
\name{ribiosPlot-package}
\alias{ribiosPlot-package}
\alias{ribiosPlot}
\title{ribiosPlot: Data transformation and visualization in ribios}
\description{
The package provides data structures and functions to plot graphics in the
ribios software suite It is a collection of code snippets I used to generate
plots to visualize omics data.
}
\author{
Jitao David Zhang <jitao_david.zhang@roche.com>
}
\keyword{package}
