% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{royalbluered}
\alias{royalbluered}
\alias{royalredblue}
\alias{royalbluegrayred}
\alias{royalredgrayblue}
\alias{turyeb}
\alias{redgreen}
\alias{greenred}
\alias{bluered}
\alias{redblue}
\alias{redblackblue}
\alias{cyanblackyellow}
\alias{yellowblackcyan}
\alias{blueblackred}
\alias{blackredyellow}
\alias{blackgoldred}
\alias{whiteblueblackheat}
\alias{heat}
\alias{magentayellow}
\alias{yellowmagenta}
\alias{blackyellow}
\alias{yellowblack}
\alias{whiteblue}
\alias{whitered}
\alias{blackred}
\alias{blackgreen}
\alias{whiteblack}
\alias{blackwhite}
\title{Two and three-color panels}
\usage{
royalbluered(n)

royalredblue(n)

royalbluegrayred(n)

royalredgrayblue(n)

blackyellow(n)

yellowblack(n)

whiteblue(n)

whitered(n)

blackred(n)

blackgreen(n)

whiteblack(n)

blackwhite(n)

turyeb(n)

redgreen(n)

greenred(n)

bluered(n)

redblue(n)

blueblackred(n)

cyanblackyellow(n)

yellowblackcyan(n)

redblackblue(n)

blackredyellow(n)

blackgoldred(n)

magentayellow(n)

yellowmagenta(n)

whiteblueblackheat(n)

heat(n)
}
\arguments{
\item{n}{Number of colors needed}
}
\value{
Character vector of length \code{n} coding colors
}
\description{
Two and three-color panels
}
\examples{

display.threecolor.panels()

}
\seealso{
\code{\link{blackyellow}} for two-color systems
}
