% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compactTrellis.R
\name{setCompactTrellis}
\alias{setCompactTrellis}
\title{Set compact trellis as default}
\usage{
setCompactTrellis()
}
\value{
as \code{lattice.options}. The side-effect is used.
}
\description{
The function sets compact trellis options as default
}
\examples{
\dontrun{
setCompactTrellis()
}
}

